modify

+++ transformer_ner_extra_different_embedding/models.py
@@ -56,7 +56,7 @@ class EncoderDecoder(nn.Module):
 
 
     def greedy_decode(self, src, src_mask, max_len, start_symbol, ner, oov_nums, vocab_size):
-        extend_mask = src < vocab_size
+        extend_mask = (src < vocab_size).long()
         memory = self.encode(src * extend_mask, src_mask, ner)
 
         #print(memory.size())
@@ -75,7 +75,7 @@ class EncoderDecoder(nn.Module):
             #print(next_word.view(-1, 1))
             ret = torch.cat([ret, 
                             (torch.zeros(src.size()[0], 1).type_as(src.data)) + (next_word.view(-1, 1))], dim=1)
-            m = next_word < vocab_size
+            m = (next_word < vocab_size).long()
             next_word = m * next_word
             ys = torch.cat([ys, 
                             (torch.zeros(src.size()[0], 1).type_as(src.data)) + (next_word.view(-1, 1))], dim=1)
